fix(branch): fixed commented issues
