+++ src/backend/api/routes/regional.py
@@ -1,4 +1,8 @@
-"""Regional Office API — AFOR Import, Regional Incidents, Stats."""
+"""Regional Office API — AFOR Import, Regional Incidents, Stats.
+
+Route handlers are grouped in sub-modules; this file registers them all
+under the /api/regional prefix.
+"""
 
 from __future__ import annotations
 
@@ -13,7 +17,6 @@ from typing import Annotated, Any, Optional
 
 from fastapi import APIRouter, Depends, File, HTTPException, Query, Request, UploadFile
 from fastapi.responses import Response
-from pydantic import BaseModel
 from sqlalchemy import text
 from sqlalchemy.orm import Session
 
@@ -48,42 +51,40 @@ from services.analytics_read_model import (
 from utils.crypto import SecurityProvider, SecurityProviderError
 from utils.audit import log_system_audit
 
+# ── Schemas extracted to schemas/regional.py ─────────────────────────────────
+from schemas.regional import (
+    RegionalStatsResponse,
+    IncidentCreateRequest,
+    IncidentUpdateRequest,
+    VerificationActionRequest,
+    CorrectionRequest,
+    BulkApproveRequest,
+)
 
-# ── Lazy SecurityProvider singleton (avoids import-time env check in test mocks) ──
-_sp_instance: SecurityProvider | None = None
+# ── Helpers extracted to services/regional_incidents/helpers.py ──────────────
+from services.regional_incidents.helpers import (
+    get_security_provider as _get_security_provider_from_helpers,
+    decrypt_pii_blob as _decrypt_pii_blob,
+    normalize_general_category as _normalize_general_category,
+    region_text_matches as _region_text_matches,
+    generate_reference_number as _generate_reference_number,
+    insert_incident_verification_history as _insert_incident_verification_history,
+    apply_incident_field_updates as _apply_incident_field_updates,
+    build_audit_log_query as _build_audit_log_query,
+    _CATEGORY_DB_VARIANTS,
+    _ivh_has_column as _incident_verification_history_has_column,
+    _ivh_uses_target_columns as _incident_verification_history_uses_target_columns,
+)
 
 
 def _get_security_provider() -> SecurityProvider:
-    global _sp_instance  # noqa: PLW0603
-    if _sp_instance is None:
-        _sp_instance = SecurityProvider()
-    return _sp_instance
+    return _get_security_provider_from_helpers()
 
 
 logger = logging.getLogger("wims.regional")
 
 
 router = APIRouter(prefix="/api/regional", tags=["regional"])
-
-
-# ---------------------------------------------------------------------------
-# Pydantic Schemas
-# ---------------------------------------------------------------------------
-
-
-class RegionalStatsResponse(BaseModel):
-    total_incidents: int
-    total_incidents_this_week: int = 0
-    by_category: list[dict[str, Any]]
-    by_alarm_level: list[dict[str, Any]]
-    by_status: list[dict[str, Any]]
-    wildland_total: int = 0
-    by_wildland_type: list[dict[str, Any]] = []
-    structures_affected: int = 0
-    households_affected: int = 0
-    families_affected: int = 0
-    individuals_affected: int = 0
-    vehicles_affected: int = 0
 
 
 def _regional_lifecycle_dependencies() -> RegionalIncidentLifecycleDependencies:
@@ -94,204 +95,10 @@ def _regional_lifecycle_dependencies() -> RegionalIncidentLifecycleDependencies:
     )
 
 
-def _incident_verification_history_uses_target_columns(db: Session) -> bool:
-    """Return True when IVH table already has target_type/target_id columns."""
-    return _incident_verification_history_has_column(db, "target_type")
-
-
-def _incident_verification_history_has_column(db: Session, column_name: str) -> bool:
-    """Return True when IVH table has the given column."""
-    return bool(
-        db.execute(
-            text("""
-                SELECT EXISTS (
-                    SELECT 1
-                    FROM information_schema.columns
-                    WHERE table_schema = 'wims'
-                      AND table_name = 'incident_verification_history'
-                      AND column_name = :column_name
-                )
-            """),
-            {"column_name": column_name},
-        ).scalar()
-    )
-
-
 def _incident_verification_history_has_hash_columns(db: Session) -> bool:
     """Return True when IVH table has columns needed for correction hash chaining."""
-    return all(
-        _incident_verification_history_has_column(db, column_name)
-        for column_name in (
-            "old_data_hash",
-            "new_data_hash",
-            "corrected_fields",
-            "prev_ivh_hash",
-            "ivh_row_hash",
-        )
-    )
-
-
-def _insert_incident_verification_history(
-    db: Session,
-    *,
-    incident_id: int,
-    actor_user_id: str,
-    previous_status: str,
-    new_status: str,
-    notes: str,
-    action_label: str | None = None,
-) -> None:
-    """Insert IVH row with compatibility for both legacy and migrated schemas."""
-    has_action_label = _incident_verification_history_has_column(db, "action_label")
-
-    if _incident_verification_history_uses_target_columns(db):
-        if has_action_label:
-            db.execute(
-                text("""
-                    INSERT INTO wims.incident_verification_history (
-                        target_type, target_id, action_by_user_id,
-                        previous_status, new_status, notes, action_label
-                    ) VALUES (
-                        'OFFICIAL', :iid, CAST(:uid AS uuid),
-                        :prev_status, :new_status, :notes, :action_label
-                    )
-                """),
-                {
-                    "iid": incident_id,
-                    "uid": actor_user_id,
-                    "prev_status": previous_status,
-                    "new_status": new_status,
-                    "notes": notes,
-                    "action_label": action_label,
-                },
-            )
-        else:
-            db.execute(
-                text("""
-                    INSERT INTO wims.incident_verification_history (
-                        target_type, target_id, action_by_user_id,
-                        previous_status, new_status, notes
-                    ) VALUES (
-                        'OFFICIAL', :iid, CAST(:uid AS uuid),
-                        :prev_status, :new_status, :notes
-                    )
-                """),
-                {
-                    "iid": incident_id,
-                    "uid": actor_user_id,
-                    "prev_status": previous_status,
-                    "new_status": new_status,
-                    "notes": notes,
-                },
-            )
-        return
-
-    db.execute(
-        text("""
-            INSERT INTO wims.incident_verification_history (
-                incident_id, action_by_user_id,
-                previous_status, new_status, comments
-            ) VALUES (
-                :iid, CAST(:uid AS uuid),
-                :prev_status, :new_status, :comments
-            )
-        """),
-        {
-            "iid": incident_id,
-            "uid": actor_user_id,
-            "prev_status": previous_status,
-            "new_status": new_status,
-            "comments": notes,
-        },
-    )
-
-
-_CATEGORY_CANONICAL: dict[str, str] = {
-    "STRUCTURAL": "STRUCTURAL",
-    "NON_STRUCTURAL": "NON_STRUCTURAL",
-    "NON-STRUCTURAL": "NON_STRUCTURAL",
-    "VEHICULAR": "TRANSPORTATION",
-    "TRANSPORTATION": "TRANSPORTATION",
-    "WILDLAND": "WILDLAND",
-}
-
-# All known DB values for each canonical category (covers legacy form submissions)
-_CATEGORY_DB_VARIANTS: dict[str, list[str]] = {
-    "STRUCTURAL": ["STRUCTURAL", "Structural"],
-    "NON_STRUCTURAL": ["NON_STRUCTURAL", "Non-Structural", "NON-STRUCTURAL"],
-    "TRANSPORTATION": ["TRANSPORTATION", "VEHICULAR", "Transportation", "Vehicular"],
-}
-
-
-def _normalize_general_category(val: str) -> str:
-    key = val.strip().upper().replace("-", "_").replace(" ", "_")
-    return _CATEGORY_CANONICAL.get(key, val)
-
-
-def _safe_int(val: Any, default: int = 0) -> int:
-    if val is None or val == "" or val == "N/A":
-        return default
-    try:
-        if isinstance(val, (int, float)):
-            return int(val)
-        return int(float(str(val).strip()))
-    except (ValueError, TypeError):
-        return default
-
-
-def _safe_float(val: Any, default: float = 0.0) -> float:
-    if val is None or val == "" or val == "N/A":
-        return default
-    try:
-        return float(str(val).strip())
-    except (ValueError, TypeError):
-        return default
-
-
-# ---------------------------------------------------------------------------
-# Region name alias matching (for AFOR import region mismatch check)
-# ---------------------------------------------------------------------------
-
-# Maps DB region_name (uppercase) → accepted alternative abbreviations/spellings.
-# The existing substring check handles most cases (e.g. "ILOCOS REGION" is already
-# in "REGION I - ILOCOS REGION"), so only add entries that substring matching misses.
-_REGION_ALIASES: dict[str, list[str]] = {
-    "NATIONAL CAPITAL REGION": ["NCR", "METRO MANILA"],
-    "CORDILLERA ADMINISTRATIVE REGION": ["CAR"],
-    "REGION I - ILOCOS REGION": ["REGION 1"],
-    "REGION II - CAGAYAN VALLEY": ["REGION 2"],
-    "REGION III - CENTRAL LUZON": ["REGION 3"],
-    "REGION IV-A - CALABARZON": ["REGION 4-A", "REGION 4A", "REGION IVA"],
-    "REGION IV-B - MIMAROPA": ["REGION 4-B", "REGION 4B", "REGION IVB"],
-    "REGION V - BICOL REGION": ["REGION 5"],
-    "REGION VI - WESTERN VISAYAS": ["REGION 6"],
-    "REGION VII - CENTRAL VISAYAS": ["REGION 7"],
-    "REGION VIII - EASTERN VISAYAS": ["REGION 8"],
-    "REGION IX - ZAMBOANGA PENINSULA": ["REGION 9"],
-    "REGION X - NORTHERN MINDANAO": ["REGION 10"],
-    "REGION XI - DAVAO REGION": ["REGION 11"],
-    "REGION XII - SOCCSKSARGEN": ["REGION 12"],
-    "REGION XIII - CARAGA": ["REGION 13"],
-    "BARMM": ["BANGSAMORO", "ARMM", "BANGSAMORO AUTONOMOUS REGION IN MUSLIM MINDANAO"],
-    "NIR - NEGROS ISLAND REGION": ["NIR", "NEGROS ISLAND REGION"],
-}
-
-
-def _region_text_matches(encoder_region_name: str, xlsx_region_text: str) -> bool:
-    """Return True if xlsx_region_text refers to the same PH region as encoder_region_name.
-
-    Handles acronym equivalents (NCR = National Capital Region, CAR = Cordillera
-    Administrative Region, etc.) and Arabic/Roman numeral variants for numbered regions.
-    """
-    enc = encoder_region_name.strip().upper()
-    xlsx = xlsx_region_text.strip().upper()
-    if enc in xlsx or xlsx in enc:
-        return True
-    for alias in _REGION_ALIASES.get(enc, []):
-        a = alias.upper()
-        if a == xlsx or a in xlsx or xlsx in a:
-            return True
-    return False
+    from services.regional_incidents.helpers import ivh_has_hash_columns
+    return ivh_has_hash_columns(db)
 
 
 # ---------------------------------------------------------------------------
@@ -524,20 +331,10 @@ def get_regional_incidents(
         caller_name = r[18]
         caller_number = r[19]
         if r[21] and r[22]:
-            try:
-                pii_plaintext = _get_security_provider().decrypt_json(
-                    r[22],
-                    r[21],
-                    f"incident_id:{r[0]}".encode("utf-8"),
-                )
-                owner_name = pii_plaintext.get("owner_name") or owner_name
-                caller_name = pii_plaintext.get("caller_name") or caller_name
-                caller_number = pii_plaintext.get("caller_number") or caller_number
-            except SecurityProviderError:
-                logger.error(
-                    "CRITICAL: PII blob decryption failed for incident list item. incident_id=%s",
-                    r[0],
-                )
+            pii_plaintext = _decrypt_pii_blob(r[22], r[21], r[0])
+            owner_name = pii_plaintext.get("owner_name") or owner_name
+            caller_name = pii_plaintext.get("caller_name") or caller_name
+            caller_number = pii_plaintext.get("caller_number") or caller_number
 
         items.append(
             {
@@ -853,28 +650,11 @@ def get_regional_incident_detail(
 
     # ── Decrypt PII blob if present (new writes use encrypted blob; old rows fall back) ──
     if sd_dict.get("pii_blob_enc") and sd_dict.get("encryption_iv"):
-        try:
-            aad = f"incident_id:{incident_id}".encode("utf-8")
-            pii_plaintext = _get_security_provider().decrypt_json(
-                sd_dict["encryption_iv"],
-                sd_dict["pii_blob_enc"],
-                aad,
-            )
-            # Inject decrypted PII fields so frontend contract is unchanged
-            sd_dict["caller_name"] = pii_plaintext.get("caller_name")
-            sd_dict["caller_number"] = pii_plaintext.get("caller_number")
-            sd_dict["owner_name"] = pii_plaintext.get("owner_name")
-            sd_dict["occupant_name"] = pii_plaintext.get("occupant_name")
-        except SecurityProviderError:
-            # Auth/key failure on a blob that claims to be valid — possible tampering
-            # or key rotation without re-encrypt. Log with incident_id; never log
-            # nonce, ciphertext, or plaintext. Return legacy plaintext as fallback.
-            logger.error(
-                "CRITICAL: PII blob decryption failed (possible tamper or key mismatch). "
-                "incident_id=%s",
-                incident_id,
-            )
-            pass
+        pii_plaintext = _decrypt_pii_blob(sd_dict["encryption_iv"], sd_dict["pii_blob_enc"], incident_id)
+        sd_dict["caller_name"] = pii_plaintext.get("caller_name")
+        sd_dict["caller_number"] = pii_plaintext.get("caller_number")
+        sd_dict["owner_name"] = pii_plaintext.get("owner_name")
+        sd_dict["occupant_name"] = pii_plaintext.get("occupant_name")
 
     # Do not expose internal blob columns in API response
     sd_dict.pop("pii_blob_enc", None)
@@ -1259,209 +1039,6 @@ def get_regional_stats(
 # CRUD — Direct Incident Create / Update / Delete
 # ---------------------------------------------------------------------------
 
-_AFOR_MONTH_CODES = [
-    "JAN",
-    "FEB",
-    "MAR",
-    "APR",
-    "MAY",
-    "JUN",
-    "JUL",
-    "AUG",
-    "SEP",
-    "OCT",
-    "NOV",
-    "DEC",
-]
-
-
-_REGION_CODE_TO_AFOR: dict[str, str] = {
-    "NCR": "RGN-NCR",
-    "CAR": "RGN-CAR",
-    "NIR": "RGN-NIR",
-    "BARMM": "RGN-BARMM",
-    "I": "RGN-1",
-    "II": "RGN-2",
-    "III": "RGN-3",
-    "IV-A": "RGN-4A",
-    "IV-B": "RGN-4B",
-    "V": "RGN-5",
-    "VI": "RGN-6",
-    "VII": "RGN-7",
-    "VIII": "RGN-8",
-    "IX": "RGN-9",
-    "X": "RGN-10",
-    "XI": "RGN-11",
-    "XII": "RGN-12",
-    "XIII": "RGN-13",
-}
-
-
-def _generate_reference_number(
-    db: Session,
-    region_id: int,
-    incident_type_code: str,
-    notification_dt: str | None,
-) -> str:
-    """Generate AFOR-{RGN-CODE}-{type}-{MMM}-{YYYY}-{NNNN}.
-
-    The sequence number is globally unique across all incidents — not per-region
-    or per-type — so no two incidents share the same trailing number.
-    """
-    region_row = db.execute(
-        text("SELECT region_code FROM wims.ref_regions WHERE region_id = :rid"),
-        {"rid": region_id},
-    ).fetchone()
-    raw_code = region_row[0] if region_row else "UNK"
-    rgn_code = _REGION_CODE_TO_AFOR.get(raw_code, f"RGN-{raw_code}")
-
-    try:
-        dt = (
-            datetime.fromisoformat(str(notification_dt).replace("Z", "+00:00"))
-            if notification_dt
-            else datetime.now()
-        )
-    except (ValueError, TypeError):
-        dt = datetime.now()
-
-    month = _AFOR_MONTH_CODES[dt.month - 1]
-    year = dt.year
-
-    # Atomic monotonic counter — persists across archive/replace flows.
-    # Fails loudly if the sequence row is missing, preventing duplicate risk.
-    seq_row = db.execute(
-        text("""
-            UPDATE wims.reference_sequence
-            SET current_value = current_value + 1
-            WHERE id = 0
-            RETURNING current_value
-        """),
-    ).fetchone()
-    if not seq_row:
-        raise RuntimeError(
-            "reference_sequence row id=0 is missing — cannot generate reference number"
-        )
-    seq = int(seq_row[0])
-    return f"AFOR-{rgn_code}-{incident_type_code}-{month}-{year}-{seq:04d}"
-
-
-class IncidentCreateRequest(BaseModel):
-    """Create a new fire incident with nonsensitive + optional sensitive details."""
-
-    latitude: float
-    longitude: float
-    region_id: int | None = None
-    # Nonsensitive details
-    notification_dt: str | None = None
-    alarm_level: str | None = None
-    general_category: str | None = None
-    sub_category: str | None = None
-    specific_type: str | None = None
-    occupancy_type: str | None = None
-    city_id: int | None = None
-    distance_from_station_km: float | None = None
-    estimated_damage_php: float | None = None
-    civilian_injured: int = 0
-    civilian_deaths: int = 0
-    firefighter_injured: int = 0
-    firefighter_deaths: int = 0
-    families_affected: int = 0
-    structures_affected: int = 0
-    households_affected: int = 0
-    individuals_affected: int = 0
-    responder_type: str | None = None
-    fire_origin: str | None = None
-    extent_of_damage: str | None = None
-    stage_of_fire: str | None = None
-    fire_station_name: str | None = None
-    total_response_time_minutes: int | None = None
-    recommendations: str | None = None
-    # Location text (free-text, replaces city_id/province join for display)
-    province_district: str | None = None
-    city_municipality: str | None = None
-    barangay: str | None = None
-    # Reference number fields
-    incident_type_code: str | None = None
-    # Update-request tracking
-    parent_incident_id: int | None = None
-    # Sensitive details (optional — PII fields)
-    street_address: str | None = None
-    landmark: str | None = None
-    caller_name: str | None = None
-    caller_number: str | None = None
-    narrative_report: str | None = None
-    owner_name: str | None = None
-    occupant_name: str | None = None
-    establishment_name: str | None = None
-    receiver_name: str | None = None
-    prepared_by_officer: str | None = None
-    noted_by_officer: str | None = None
-    remarks: str | None = None
-
-
-class IncidentUpdateRequest(BaseModel):
-    """Update an existing DRAFT/PENDING incident."""
-
-    # Nonsensitive fields
-    notification_dt: str | None = None
-    alarm_level: str | None = None
-    general_category: str | None = None
-    sub_category: str | None = None
-    specific_type: str | None = None
-    occupancy_type: str | None = None
-    city_id: int | None = None
-    distance_from_station_km: float | None = None
-    estimated_damage_php: float | None = None
-    civilian_injured: int | None = None
-    civilian_deaths: int | None = None
-    firefighter_injured: int | None = None
-    firefighter_deaths: int | None = None
-    families_affected: int | None = None
-    structures_affected: int | None = None
-    households_affected: int | None = None
-    individuals_affected: int | None = None
-    responder_type: str | None = None
-    fire_origin: str | None = None
-    extent_of_damage: str | None = None
-    extent_total_floor_area_sqm: float | None = None
-    extent_total_land_area_hectares: float | None = None
-    stage_of_fire: str | None = None
-    general_description_of_involved: str | None = None
-    fire_station_name: str | None = None
-    total_response_time_minutes: int | None = None
-    vehicles_affected: int | None = None
-    recommendations: str | None = None
-    # Location text (free-text, replaces city_id/province join for display)
-    province_district: str | None = None
-    city_municipality: str | None = None
-    barangay: str | None = None
-    # Reference number fields
-    incident_type_code: str | None = None
-    # Sensitive fields
-    street_address: str | None = None
-    landmark: str | None = None
-    caller_name: str | None = None
-    caller_number: str | None = None
-    narrative_report: str | None = None
-    owner_name: str | None = None
-    occupant_name: str | None = None
-    establishment_name: str | None = None
-    receiver_name: str | None = None
-    prepared_by_officer: str | None = None
-    noted_by_officer: str | None = None
-    remarks: str | None = None
-    # JSONB fields for full-form edit
-    alarm_timeline: dict | None = None
-    resources_deployed: dict | None = None
-    problems_encountered: list | None = None
-    other_personnel: list | None = None
-    personnel_on_duty: dict | None = None
-    casualty_details: dict | None = None
-    disposition: str | None = None
-    latitude: float | None = None
-    longitude: float | None = None
-
-
 @router.post("/incidents", status_code=201)
 def create_incident(
     body: IncidentCreateRequest,
@@ -1633,228 +1210,6 @@ def create_incident(
         "incident_type_code": type_code,
         "parent_incident_id": body.parent_incident_id,
     }
-
-
-def _apply_incident_field_updates(
-    db: Session, incident_id: int, body: "IncidentUpdateRequest"
-) -> None:
-    """Apply nonsensitive/sensitive/JSONB/coords field updates from an
-    IncidentUpdateRequest to the given incident_id. Caller is responsible
-    for status checks, audit-trail writes, and committing the transaction.
-    """
-    # Ensure child rows exist so UPDATE statements never silently affect 0 rows.
-    db.execute(
-        text(
-            """
-            INSERT INTO wims.incident_nonsensitive_details (incident_id)
-            SELECT :iid
-            WHERE NOT EXISTS (
-                SELECT 1 FROM wims.incident_nonsensitive_details WHERE incident_id = :iid
-            )
-            """
-        ),
-        {"iid": incident_id},
-    )
-    db.execute(
-        text(
-            """
-            INSERT INTO wims.incident_sensitive_details (incident_id)
-            SELECT :iid
-            WHERE NOT EXISTS (
-                SELECT 1 FROM wims.incident_sensitive_details WHERE incident_id = :iid
-            )
-            """
-        ),
-        {"iid": incident_id},
-    )
-
-    ns_fields = {
-        "notification_dt",
-        "alarm_level",
-        "general_category",
-        "sub_category",
-        "specific_type",
-        "occupancy_type",
-        "city_id",
-        "barangay_id",
-        "province_district",
-        "city_municipality",
-        "barangay",
-        "distance_from_station_km",
-        "estimated_damage_php",
-        "civilian_injured",
-        "civilian_deaths",
-        "firefighter_injured",
-        "firefighter_deaths",
-        "families_affected",
-        "structures_affected",
-        "households_affected",
-        "individuals_affected",
-        "responder_type",
-        "fire_origin",
-        "extent_of_damage",
-        "extent_total_floor_area_sqm",
-        "extent_total_land_area_hectares",
-        "stage_of_fire",
-        "general_description_of_involved",
-        "fire_station_name",
-        "total_response_time_minutes",
-        "vehicles_affected",
-        "recommendations",
-    }
-    ns_updates: list[str] = []
-    ns_params: dict[str, Any] = {"iid": incident_id}
-    for field in ns_fields:
-        val = getattr(body, field, None)
-        if val is not None:
-            if field == "alarm_level" and isinstance(val, str):
-                val = ALARM_LEVEL_MAP.get(val.upper().strip(), val)
-            elif field == "general_category" and isinstance(val, str):
-                val = _normalize_general_category(val)
-            ns_updates.append(f"{field} = :{field}")
-            ns_params[field] = val
-    if ns_updates:
-        db.execute(
-            text(
-                f"UPDATE wims.incident_nonsensitive_details SET {', '.join(ns_updates)} WHERE incident_id = :iid"
-            ),
-            ns_params,
-        )
-
-    # Update incident_type_code on the fire_incidents core row if provided
-    new_type_code = (getattr(body, "incident_type_code", None) or "").strip().upper() or None
-    if new_type_code:
-        db.execute(
-            text(
-                "UPDATE wims.fire_incidents SET incident_type_code = :tc WHERE incident_id = :iid"
-            ),
-            {"tc": new_type_code, "iid": incident_id},
-        )
-
-    sd_fields = {
-        "street_address",
-        "landmark",
-        "narrative_report",
-        "establishment_name",
-        "receiver_name",
-        "prepared_by_officer",
-        "noted_by_officer",
-        "remarks",
-    }
-    pii_fields = ["caller_name", "caller_number", "owner_name", "occupant_name"]
-    sd_updates: list[str] = []
-    sd_params: dict[str, Any] = {"iid": incident_id}
-    has_pii_update = False
-    for field in sd_fields | set(pii_fields):
-        val = getattr(body, field, None)
-        if val is not None:
-            if field in pii_fields:
-                has_pii_update = True
-                # owner_name also mirrors to the plaintext column (used by list queries)
-                if field == "owner_name":
-                    sd_updates.append(f"{field} = :{field}")
-                    sd_params[field] = val
-            else:
-                sd_updates.append(f"{field} = :{field}")
-                sd_params[field] = val
-    if has_pii_update:
-        existing = db.execute(
-            text(
-                "SELECT pii_blob_enc, encryption_iv FROM wims.incident_sensitive_details WHERE incident_id = :iid"
-            ),
-            {"iid": incident_id},
-        ).fetchone()
-        existing_pii: dict[str, Any] = {}
-        if existing and existing[0] and existing[1]:
-            try:
-                sp = _get_security_provider()
-                existing_pii = sp.decrypt_json(
-                    existing[1], existing[0], f"incident_id:{incident_id}".encode()
-                )
-            except SecurityProviderError:
-                logger.warning(
-                    "Failed to decrypt existing PII for incident %s — overwriting",
-                    incident_id,
-                )
-        for field in pii_fields:
-            val = getattr(body, field, None)
-            if val is not None:
-                existing_pii[field] = val
-        try:
-            sp = _get_security_provider()
-            nonce_b64, ct_b64 = sp.encrypt_json(existing_pii, f"incident_id:{incident_id}".encode())
-            sd_updates.extend(["pii_blob_enc = :pii_blob", "encryption_iv = :enc_iv"])
-            sd_params["pii_blob"] = ct_b64
-            sd_params["enc_iv"] = nonce_b64
-        except SecurityProviderError:
-            logger.warning("PII re-encryption failed for incident %s", incident_id)
-    if sd_updates:
-        db.execute(
-            text(
-                f"UPDATE wims.incident_sensitive_details SET {', '.join(sd_updates)} WHERE incident_id = :iid"
-            ),
-            sd_params,
-        )
-
-    jsonb_ns = {
-        "alarm_timeline": body.alarm_timeline,
-        "resources_deployed": body.resources_deployed,
-        "problems_encountered": body.problems_encountered,
-    }
-    jsonb_ns_updates: list[str] = []
-    jsonb_ns_params: dict[str, Any] = {"iid": incident_id}
-    for field, val in jsonb_ns.items():
-        if val is not None:
-            jsonb_ns_updates.append(f"{field} = CAST(:{field} AS jsonb)")
-            jsonb_ns_params[field] = json.dumps(val)
-    if jsonb_ns_updates:
-        db.execute(
-            text(
-                f"UPDATE wims.incident_nonsensitive_details SET {', '.join(jsonb_ns_updates)} WHERE incident_id = :iid"
-            ),
-            jsonb_ns_params,
-        )
-
-    jsonb_sd = {
-        "personnel_on_duty": body.personnel_on_duty,
-        "other_personnel": body.other_personnel,
-        "casualty_details": body.casualty_details,
-        "disposition": body.disposition,
-    }
-    jsonb_sd_updates: list[str] = []
-    jsonb_sd_params: dict[str, Any] = {"iid": incident_id}
-    for field, val in jsonb_sd.items():
-        if val is not None:
-            if field == "disposition":
-                jsonb_sd_updates.append(f"{field} = :{field}")
-            else:
-                jsonb_sd_updates.append(f"{field} = CAST(:{field} AS jsonb)")
-            jsonb_sd_params[field] = json.dumps(val) if field != "disposition" else val
-    if jsonb_sd_updates:
-        db.execute(
-            text(
-                f"UPDATE wims.incident_sensitive_details SET {', '.join(jsonb_sd_updates)} WHERE incident_id = :iid"
-            ),
-            jsonb_sd_params,
-        )
-
-    if body.latitude is not None and body.longitude is not None:
-        db.execute(
-            text(
-                """
-                UPDATE wims.fire_incidents
-                SET updated_at = now(),
-                    location = ST_SetSRID(ST_MakePoint(:lon, :lat), 4326)
-                WHERE incident_id = :iid
-                """
-            ),
-            {"lon": body.longitude, "lat": body.latitude, "iid": incident_id},
-        )
-    else:
-        db.execute(
-            text("UPDATE wims.fire_incidents SET updated_at = now() WHERE incident_id = :iid"),
-            {"iid": incident_id},
-        )
 
 
 @router.put("/incidents/{incident_id}")
@@ -2100,26 +1455,6 @@ def submit_incident_for_review(
 # ---------------------------------------------------------------------------
 # Validator Workflow
 # ---------------------------------------------------------------------------
-
-
-# Allowed actions a NATIONAL_VALIDATOR can submit and their target DB status.
-# accept_replace: approve a duplicate by inheriting the matched incident's ref_num and archiving it.
-class VerificationActionRequest(BaseModel):
-    """Body for PATCH /api/regional/incidents/{incident_id}/verification."""
-
-    action: str  # "accept" | "accept_replace" | "pending" | "reject"
-    notes: str | None = None  # Optional reason / validator notes
-    # When the validator chooses "Replace Existing" from the duplicate modal, pass the ID
-    # of the incident to supersede. Takes priority over the stored duplicate_of value.
-    original_incident_id: int | None = None
-
-
-class CorrectionRequest(BaseModel):
-    """Body for PATCH /api/regional/incidents/{incident_id}/correct."""
-
-    corrections: dict
-    notes: str | None = None
-
 
 @router.get("/validator/incidents")
 def get_validator_incident_queue(
@@ -2555,11 +1890,6 @@ async def correct_verified_incident(
 # ---------------------------------------------------------------------------
 
 
-class BulkApproveRequest(BaseModel):
-    incident_ids: list[int]
-    notes: str | None = None
-
-
 @router.post("/validator/incidents/bulk-approve")
 def bulk_approve_incidents(
     body: BulkApproveRequest,
@@ -2852,42 +2182,6 @@ def get_encoder_audit_log(
         "limit": limit,
         "offset": offset,
     }
-
-
-def _build_audit_log_query(
-    *,
-    date_from: str | None,
-    date_to: str | None,
-    region_id: int | None,
-    actor_username: str | None,
-    role: str | None,
-    action: str | None,
-) -> tuple[str, dict[str, Any]]:
-    """Compose a parameterized WHERE clause for audit log queries.
-
-    Returns (where_sql, params). The caller plugs where_sql into a SELECT.
-    """
-    where_clauses = ["ivh.target_type = 'OFFICIAL'"]
-    params: dict[str, Any] = {}
-    if date_from:
-        where_clauses.append("ivh.action_timestamp >= CAST(:date_from AS timestamptz)")
-        params["date_from"] = date_from
-    if date_to:
-        where_clauses.append("ivh.action_timestamp <= CAST(:date_to AS timestamptz)")
-        params["date_to"] = date_to
-    if region_id is not None:
-        where_clauses.append("fi.region_id = :region_id")
-        params["region_id"] = region_id
-    if actor_username:
-        where_clauses.append("u.username ILIKE :actor_username")
-        params["actor_username"] = f"%{actor_username}%"
-    if role:
-        where_clauses.append("u.role = :role")
-        params["role"] = role
-    if action:
-        where_clauses.append("ivh.action_label = :action")
-        params["action"] = action
-    return " AND ".join(where_clauses), params
 
 
 @router.get("/validator/audit-logs")
