+++ src/backend/schemas/regional.py
@@ -1,0 +1,149 @@
+"""Pydantic schemas for the Regional Encoder / National Validator API."""
+
+from __future__ import annotations
+
+from typing import Any
+
+from pydantic import BaseModel
+
+
+class RegionalStatsResponse(BaseModel):
+    total_incidents: int
+    total_incidents_this_week: int = 0
+    by_category: list[dict[str, Any]]
+    by_alarm_level: list[dict[str, Any]]
+    by_status: list[dict[str, Any]]
+    wildland_total: int = 0
+    by_wildland_type: list[dict[str, Any]] = []
+    structures_affected: int = 0
+    households_affected: int = 0
+    families_affected: int = 0
+    individuals_affected: int = 0
+    vehicles_affected: int = 0
+
+
+class IncidentCreateRequest(BaseModel):
+    """Create a new fire incident with nonsensitive + optional sensitive details."""
+
+    latitude: float
+    longitude: float
+    region_id: int | None = None
+    notification_dt: str | None = None
+    alarm_level: str | None = None
+    general_category: str | None = None
+    sub_category: str | None = None
+    specific_type: str | None = None
+    occupancy_type: str | None = None
+    city_id: int | None = None
+    distance_from_station_km: float | None = None
+    estimated_damage_php: float | None = None
+    civilian_injured: int = 0
+    civilian_deaths: int = 0
+    firefighter_injured: int = 0
+    firefighter_deaths: int = 0
+    families_affected: int = 0
+    structures_affected: int = 0
+    households_affected: int = 0
+    individuals_affected: int = 0
+    responder_type: str | None = None
+    fire_origin: str | None = None
+    extent_of_damage: str | None = None
+    stage_of_fire: str | None = None
+    fire_station_name: str | None = None
+    total_response_time_minutes: int | None = None
+    recommendations: str | None = None
+    province_district: str | None = None
+    city_municipality: str | None = None
+    barangay: str | None = None
+    incident_type_code: str | None = None
+    parent_incident_id: int | None = None
+    street_address: str | None = None
+    landmark: str | None = None
+    caller_name: str | None = None
+    caller_number: str | None = None
+    narrative_report: str | None = None
+    owner_name: str | None = None
+    occupant_name: str | None = None
+    establishment_name: str | None = None
+    receiver_name: str | None = None
+    prepared_by_officer: str | None = None
+    noted_by_officer: str | None = None
+    remarks: str | None = None
+
+
+class IncidentUpdateRequest(BaseModel):
+    """Update an existing DRAFT/PENDING incident."""
+
+    notification_dt: str | None = None
+    alarm_level: str | None = None
+    general_category: str | None = None
+    sub_category: str | None = None
+    specific_type: str | None = None
+    occupancy_type: str | None = None
+    city_id: int | None = None
+    distance_from_station_km: float | None = None
+    estimated_damage_php: float | None = None
+    civilian_injured: int | None = None
+    civilian_deaths: int | None = None
+    firefighter_injured: int | None = None
+    firefighter_deaths: int | None = None
+    families_affected: int | None = None
+    structures_affected: int | None = None
+    households_affected: int | None = None
+    individuals_affected: int | None = None
+    responder_type: str | None = None
+    fire_origin: str | None = None
+    extent_of_damage: str | None = None
+    extent_total_floor_area_sqm: float | None = None
+    extent_total_land_area_hectares: float | None = None
+    stage_of_fire: str | None = None
+    general_description_of_involved: str | None = None
+    fire_station_name: str | None = None
+    total_response_time_minutes: int | None = None
+    vehicles_affected: int | None = None
+    recommendations: str | None = None
+    province_district: str | None = None
+    city_municipality: str | None = None
+    barangay: str | None = None
+    incident_type_code: str | None = None
+    street_address: str | None = None
+    landmark: str | None = None
+    caller_name: str | None = None
+    caller_number: str | None = None
+    narrative_report: str | None = None
+    owner_name: str | None = None
+    occupant_name: str | None = None
+    establishment_name: str | None = None
+    receiver_name: str | None = None
+    prepared_by_officer: str | None = None
+    noted_by_officer: str | None = None
+    remarks: str | None = None
+    alarm_timeline: dict | None = None
+    resources_deployed: dict | None = None
+    problems_encountered: list | None = None
+    other_personnel: list | None = None
+    personnel_on_duty: dict | None = None
+    casualty_details: dict | None = None
+    disposition: str | None = None
+    latitude: float | None = None
+    longitude: float | None = None
+
+
+class VerificationActionRequest(BaseModel):
+    """Body for PATCH /api/regional/incidents/{incident_id}/verification."""
+
+    action: str  # "accept" | "accept_replace" | "pending" | "reject"
+    notes: str | None = None
+    original_incident_id: int | None = None  # For accept_replace: ID to supersede
+
+
+class CorrectionRequest(BaseModel):
+    """Body for PATCH /api/regional/incidents/{incident_id}/correct."""
+
+    corrections: dict
+    notes: str | None = None
+
+
+class BulkApproveRequest(BaseModel):
+    incident_ids: list[int]
+    notes: str | None = None
