+++ src/backend/services/regional_incidents/lifecycle.py
@@ -287,7 +287,6 @@ def submit_incident_for_review_command(
                 db,
                 incident_id=incident_id,
                 region_id=geo_meta[3],
-                alarm_level=geo_meta[4],
                 incident_date=_date_str(geo_meta[0]),
                 notification_dt=geo_meta[0],
                 lat=geo_meta[5],
@@ -338,7 +337,6 @@ def submit_incident_for_review_command(
                     db,
                     incident_id=incident_id,
                     region_id=geo_meta[3],
-                    alarm_level=geo_meta[4],
                     incident_date=_date_str(geo_meta[0]),
                     notification_dt=geo_meta[0],
                     lat=geo_meta[5],
@@ -511,7 +509,6 @@ def verify_incident_command(
                 db,
                 incident_id=incident_id,
                 region_id=geo_row[5],
-                alarm_level=geo_row[6],
                 incident_date=_date_str(geo_row[2]),
                 notification_dt=geo_row[2],
                 lat=geo_row[0],
@@ -756,7 +753,6 @@ def bulk_approve_pending_incidents(
                 db,
                 incident_id=iid,
                 region_id=region_id,
-                alarm_level=alarm,
                 incident_date=_date_str(notif_dt),
                 notification_dt=notif_dt,
                 lat=lat,
