+++ src/frontend/src/components/ui/index.ts
@@ -1,0 +1,4 @@
+export { StatusBadge, STATUS_COLORS, STATUS_LABELS } from './StatusBadge';
+export { MetricPill } from './MetricPill';
+export { FilterChips } from './FilterChips';
+export { PaginationControls } from './PaginationControls';
